optimize state and add persistent mapping(OpenGL 4.4)

+++ include/caravangl_arg_indices.h
@@ -9,7 +9,7 @@
 #define SCHEMA_INIT(X) X(IDX_INIT_LOADER, "loader", PyObject *, 1)
 
 #define SCHEMA_BUF_INIT(X)                                                                         \
-    X(IDX_BUF_SIZE, "size", int, 1)                                                                \
+    X(IDX_BUF_SIZE, "size", Py_ssize_t, 1)                                                                \
     X(IDX_BUF_DATA, "data", PyObject *, 0)                                                         \
     X(IDX_BUF_TARGET, "target", uint32_t, 0)                                                       \
     X(IDX_BUF_USAGE, "usage", uint32_t, 0)
